Fix LP64 jaudio heap and Acmd pointer transport

+++ src/static/jaudio_NES/internal/driver.c
@@ -43,7 +43,7 @@ static u32 Env_Data_L3 = 0x90AAC4DE;
 
 static Acmd* __LoadAuxBuf(Acmd* cmd, u16 ofs, u16 startPos, s32 size, delay* del_p);
 static Acmd* __SaveAuxBuf(Acmd* cmd, u16 ofs, u16 startPos, s32 size, delay* del_p);
-static Acmd* Nas_SaveBufferAuto(Acmd* cmd, u16 ofs, u16 size, s32 startAddr);
+static Acmd* Nas_SaveBufferAuto(Acmd* cmd, u16 ofs, u16 size, uintptr_t startAddr);
 
 static void Nas_CpuFX(s32 chunkLen, s32 updateIdx, s32 reverbIdx) {
     delayparam* param_p;
@@ -272,11 +272,11 @@ static void Nas_SetEnvParam(Acmd* cmd, s32 reverbVol, s32 rampReverb, s32 rampL,
     aSetEnvParam(cmd, reverbVol, rampReverb, rampL, rampR);
 }
 
-static void Nas_LoadBuffer2(Acmd* cmd, s32 dst, s32 len, s32 src) {
+static void Nas_LoadBuffer2(Acmd* cmd, s32 dst, s32 len, uintptr_t src) {
     aLoadBuffer2(cmd, src, dst, len);
 }
 
-static void Nas_SaveBuffer2(Acmd* cmd, s32 src, s32 len, s32 dst) {
+static void Nas_SaveBuffer2(Acmd* cmd, s32 src, s32 len, uintptr_t dst) {
     aSaveBuffer2(cmd, dst, src, len);
 }
 
@@ -284,7 +284,7 @@ static void Nas_SetEnvParam2(Acmd* cmd, s32 volL, s32 volR) {
     aSetEnvParam2(cmd, volL, volR);
 }
 
-static void Nas_PCM8dec(Acmd* cmd, s32 flags, s32 state) {
+static void Nas_PCM8dec(Acmd* cmd, s32 flags, uintptr_t state) {
     aPCM8dec(cmd, flags, state);
 }
 
@@ -338,17 +338,17 @@ static Acmd* Nas_SaveAuxBufferCH(Acmd* cmd, delay* del_p, s16 update_idx) {
     aDMEMMove(cmd++, 0xC40, 0x3A0, size);
     aSetBuffer(cmd++, 0, 0x3A0, 0x6E0, param->save_resample_num_samples * SAMPLE_SIZE);
     aResample(cmd++, del_p->resample_flags, param->save_resample_pitch, del_p->left_save_resample_buf);
-    cmd = Nas_SaveBufferAuto(cmd, 0x6E0, param->size, (s32)&del_p->left_reverb_buf[param->start_pos]);
+    cmd = Nas_SaveBufferAuto(cmd, 0x6E0, param->size, (uintptr_t)&del_p->left_reverb_buf[param->start_pos]);
     if (param->wrapped_size != 0) {
-        cmd = Nas_SaveBufferAuto(cmd, 0x6E0 + param->size, param->wrapped_size, (s32)&del_p->left_reverb_buf[0]);
+        cmd = Nas_SaveBufferAuto(cmd, 0x6E0 + param->size, param->wrapped_size, (uintptr_t)&del_p->left_reverb_buf[0]);
     }
 
     aDMEMMove(cmd++, 0xDE0, 0x3A0, size);
     aSetBuffer(cmd++, 0, 0x3A0, 0x6E0, param->save_resample_num_samples * SAMPLE_SIZE);
     aResample(cmd++, del_p->resample_flags, param->save_resample_pitch, del_p->right_save_resample_buf);
-    cmd = Nas_SaveBufferAuto(cmd, 0x6E0, param->size, (s32)&del_p->right_reverb_buf[param->start_pos]);
+    cmd = Nas_SaveBufferAuto(cmd, 0x6E0, param->size, (uintptr_t)&del_p->right_reverb_buf[param->start_pos]);
     if (param->wrapped_size != 0) {
-        cmd = Nas_SaveBufferAuto(cmd, 0x6E0 + param->size, param->wrapped_size, (s32)&del_p->right_reverb_buf[0]);
+        cmd = Nas_SaveBufferAuto(cmd, 0x6E0 + param->size, param->wrapped_size, (uintptr_t)&del_p->right_reverb_buf[0]);
     }
 
     return cmd;
@@ -428,10 +428,10 @@ static Acmd* Nas_LoadAuxBuffer1_B(Acmd* cmd, s32 samples_per_update, delay* del_
     return cmd;
 }
 
-static Acmd* Nas_SaveBufferAuto(Acmd* cmd, u16 dmem, u16 size, s32 startAddr) {
-    s32 startUnaligned = startAddr & 15;
-    s32 endAddr = startAddr + size;
-    s32 endUnaligned = endAddr & 15;
+static Acmd* Nas_SaveBufferAuto(Acmd* cmd, u16 dmem, u16 size, uintptr_t startAddr) {
+    uintptr_t startUnaligned = startAddr & 15u;
+    uintptr_t endAddr = startAddr + size;
+    uintptr_t endUnaligned = endAddr & 15u;
 
     if (endUnaligned != 0) {
         aLoadBuffer2(cmd++, (endAddr - endUnaligned), 0x380, 16);
@@ -515,13 +515,13 @@ static Acmd* Nas_SaveAuxBuffer(Acmd* cmd, delay* del_p, s16 update_idx) {
             }
 
             if (param->size != 0) {
-                cmd = Nas_SaveBufferAuto(cmd, 0xC40, (u16)param->size, (s32)&del_p->left_reverb_buf[param->start_pos]);
-                cmd = Nas_SaveBufferAuto(cmd, 0xDE0, (u16)param->size, (s32)&del_p->right_reverb_buf[param->start_pos]);
+                cmd = Nas_SaveBufferAuto(cmd, 0xC40, (u16)param->size, (uintptr_t)&del_p->left_reverb_buf[param->start_pos]);
+                cmd = Nas_SaveBufferAuto(cmd, 0xDE0, (u16)param->size, (uintptr_t)&del_p->right_reverb_buf[param->start_pos]);
             }
 
             if (param->wrapped_size != 0) {
-                cmd = Nas_SaveBufferAuto(cmd, 0xC40 + param->size, (u16)param->wrapped_size, (s32)&del_p->left_reverb_buf[0]);
-                cmd = Nas_SaveBufferAuto(cmd, 0xDE0 + param->size, (u16)param->wrapped_size, (s32)&del_p->right_reverb_buf[0]);
+                cmd = Nas_SaveBufferAuto(cmd, 0xC40 + param->size, (u16)param->wrapped_size, (uintptr_t)&del_p->left_reverb_buf[0]);
+                cmd = Nas_SaveBufferAuto(cmd, 0xDE0 + param->size, (u16)param->wrapped_size, (uintptr_t)&del_p->right_reverb_buf[0]);
             }
             break;
     }
@@ -928,7 +928,7 @@ extern Acmd* Nas_SynthMain(s32 chan_id, commonch* common, driverch* driver, s16*
                         } else {
                             Nas_LoadBuffer2(cmd++, DMEM_UNCOMPRESSED_NOTE,
                                 numSamplesToLoadAdj * SAMPLE_SIZE + SAMPLES_PER_FRAME * SAMPLE_SIZE,
-                                (s32)reverbAddrSrc);
+                                (uintptr_t)reverbAddrSrc);
                             // deviation from MM:
                             goto codec_continue_and_skip;
                             // flags = A_CONTINUE;
@@ -1054,7 +1054,7 @@ codec_continue_and_skip:
                         Nas_SetBuffer(cmd++, 0, sampleDataDmemAddr + sampleDataChunkAlignPad,
                                                 DMEM_UNCOMPRESSED_NOTE + dmemUncompressedAddrOffset2,
                                                 numSamplesToDecode * SAMPLE_SIZE);
-                        Nas_PCM8dec(cmd++, flags, (s32)driver->synth_params->adpcm_state);
+                        Nas_PCM8dec(cmd++, flags, (uintptr_t)driver->synth_params->adpcm_state);
                         break;
 
                     case CODEC_UNK7:
@@ -1186,7 +1186,7 @@ codec_continue_and_skip:
         if (!STOP_VELOCONV) {
             // Load the velocity convolution table into DMEM_0x800
             s32 vel_conv_idx = driver->vel_conv_table_idx;
-            aLoadBuffer2(cmd++, (u32)VELOCONV_TABLE[vel_conv_idx], 0x800, sizeof(VELOCONV_TABLE[vel_conv_idx]));
+            aLoadBuffer2(cmd++, VELOCONV_TABLE[vel_conv_idx], 0x800, sizeof(VELOCONV_TABLE[vel_conv_idx]));
             aUnkCmd3(cmd++, DMEM_TEMP, 0x800, samples_per_update);
         }
 
@@ -1209,9 +1209,9 @@ codec_continue_and_skip:
                 Nas_ClearBuffer(cmd++, DMEM_COMB_TEMP - combFilterSize, combFilterSize);
                 driver->comb_filter_needs_init = false;
             } else {
-                Nas_LoadBuffer2(cmd++, DMEM_COMB_TEMP - combFilterSize, combFilterSize, (s32)combFilterState);
+                Nas_LoadBuffer2(cmd++, DMEM_COMB_TEMP - combFilterSize, combFilterSize, (uintptr_t)combFilterState);
             }
-            Nas_SaveBuffer2(cmd++, DMEM_TEMP - combFilterSize + size, combFilterSize, (s32)combFilterState);
+            Nas_SaveBuffer2(cmd++, DMEM_TEMP - combFilterSize + size, combFilterSize, (uintptr_t)combFilterState);
             Nas_Mix(cmd++, size >> 4, combFilterGain, DMEM_COMB_TEMP, DMEM_COMB_TEMP - combFilterSize);
             Nas_DMEMMove(cmd++, DMEM_COMB_TEMP - combFilterSize, DMEM_TEMP, size);
         } else {
